Separate the handler from the runnable program
